Move LongCycle to box tests
